fix(recipes): print recipes even if local storage is empty
When local storage was empty, the recipes was saved successfully but
the recipes list was not displayed on screen without refreshing the
browser.
So I moved the logic to save the recipes from AppComponent to
RecipesListComponent to avoid some repeat and I display recipes from
API if local storage was empty before.